desloppify: organize CodeScope.App root into concern folders
CodeScope.App's root had 8 .cs files mixing pollers, persistence
stores, update wiring, and the actual composition entry. Split into
3 concern folders so the root only carries App.xaml.cs and
MainWindow.xaml.cs (the composition entry):

- Polling/   ← PollBackoff, WorktreePoller, WorktreeStatusPoller,
               PullRequestStatusPoller
- Persistence/ ← LayoutStore, WindowGeometryStore
- Updates/   ← UpdateService, VersionInfo

Namespaces updated to NoScope.CodeScope.App.{Polling,Persistence,
Updates}. Consumers (App.xaml.cs, MainWindow.xaml/.cs) updated;
MainWindow.xaml gets an `upd:` xmlns for VersionInfo.Display.

Also drop null! on MainViewModel.Sidebar (type_safety review fix
folded into this batch — Sidebar is now SidebarViewModel? with a
doc-comment lifecycle note).
